THREDDS: Refactoring.
